requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
python-dotenv==0.19.0
google-generativeai==0.1.0
diagrams==0.23.3
graphviz==0.20.1
streamlit==1.22.0
pillow==9.5.0
requests==2.31.0
gunicorn==20.1.0
psycopg2-binary==2.9.3 